FIX: this fixes a bug where the model casted  to integer causing issuess

+++ src/main/java/ch/unibas/dmi/dbis/cs108/example/gui/javafx/mvc/model/GameModel.java
@@ -89,12 +89,12 @@ public class GameModel {
             .orElse(null);
 
     if (player != null) {
-      player.xPosition().set((int) x); // or double if you change your Player class
-      player.yPosition().set((int) y);
+      player.xPosition().set(x);
+      player.yPosition().set(y);
       player.setScore(score);
       // Note: You could also add a roleProperty to the Player class
     } else {
-      lobbyPlayers.add(new Player(name, role, score, (int) x, (int) y));
+      lobbyPlayers.add(new Player(name, role, score, x, y));
     }
   }
   private void loadMaps() {
